Rename: NewBadge -> DismissableBadge
