+++ app/components-react/shared/DismissableBadge.m.less
@@ -2,7 +2,7 @@
 @import '../../styles/mixins';
 @import '../../styles/badges';
 
-.new-badge {
+.dismissable-badge {
   background: var(--new-badge-text);
   color: var(--titlebar);
   font-size: 9px;
